Everything is broken (see description)
In an attempt to add "bars" to each basket (stationary rigid circles on each side of the basket) MyGame no longer loads correctly. I believe this is because, at the time of the creation of the baskets the constant color shader has not yet loaded. See line 23 of Renderable.js. When the first basket is loaded, each bar outputs null for their respective Renderable.mShader. Not sure how to fix this so I'm pushing and going to bed